remove lightningGraph and update GraphContainer
