+++ minigu/context/Cargo.toml
@@ -11,6 +11,7 @@ miette = { workspace = true }
 minigu-catalog = { workspace = true }
 minigu-common = { workspace = true }
 minigu-storage = { workspace = true }
+minigu-transaction = { workspace = true }
 rayon = { workspace = true }
 thiserror = { workspace = true }
