+++ minigu/transaction/src/lib.rs
@@ -1,7 +1,5 @@
-pub mod lightning_graph;
 pub mod transaction;
 
-pub use lightning_graph::LightningGraph;
 pub use transaction::{
     CatalogTxnState, GraphTxnState, Transaction, TransactionCore, TxnError, TxnResult, TxnState,
 };
